+++ Gruntfile.js
@@ -21,6 +21,15 @@ module.exports = function(grunt) {
     pkg: grunt.file.readJSON('bower.json'),
 
     /**
+     * Set some src and dist location variables.
+     */
+    loc: {
+      src: './src',
+      dist: '.',
+      test: './_tests'
+    },
+
+    /**
      * Bower: https://github.com/yatskevich/grunt-bower-task
      *
      * Install Bower packages and migrate static assets.
@@ -52,47 +61,6 @@ module.exports = function(grunt) {
      * Replace strings on files by using string or regex patters.
      */
     'string-replace': {
-      'static-legacy': {
-        files: {
-          'static-legacy/css/styles.css': 'static-legacy/css/styles.css'
-        },
-        options: {
-          replacements: [
-            {
-              pattern: /"Avenir Next"/ig,
-              replacement: '"AvenirNextLTW01-Regular"'
-            },
-            {
-              pattern: /"Avenir Next Regular"/ig,
-              replacement: '"AvenirNextLTW01-Regular"'
-            },
-            {
-              pattern: /"Avenir Next Demi"/ig,
-              replacement: '"AvenirNextLTW01-Demi"'
-            },
-            {
-              pattern: /"Avenir Next Demi Italic"/ig,
-              replacement: '"AvenirNextLTW01-Demi"'
-            },
-            {
-              pattern: /"Avenir Next Medium"/ig,
-              replacement: '"AvenirNextLTW01-Medium"'
-            },
-            {
-              pattern: /font-weight:(?:\s)*400;/ig,
-              replacement: 'font-family: "AvenirNextLTW01-Regular", Arial, sans-serif;'
-            },
-            {
-              pattern: /font-weight:(?:\s)*500;/ig,
-              replacement: 'font-family: "AvenirNextLTW01-Medium", Arial, sans-serif;'
-            },
-            {
-              pattern: /font-weight:(?:\s)*600;/ig,
-              replacement: 'font-family: "AvenirNextLTW01-Demi", Arial, sans-serif;'
-            }
-          ]
-        }
-      },
       chosen: {
         files: {
           'vendor/chosen/': 'vendor/chosen/chosen.css'
@@ -138,11 +106,13 @@ module.exports = function(grunt) {
           paths: grunt.file.expand('vendor/**/'),
           compress: true,
           sourceMap: true,
-          sourceMapFilename: 'static/css/main.css.map', // where file is generated and located
-          sourceMapURL: 'main.css.map', // the complete url and filename put in the compiled css file
+          // Where the sourcemap file is generated and located.
+          sourceMapFilename: '<%= loc.dist %>/static/css/main.css.map',
+          // The complete URL and sourcemap filename put in the compiled CSS file.
+          sourceMapURL: 'main.css.map',
         },
         files: {
-          'static/css/main.css': ['static/css/main.less']
+          '<%= loc.dist %>/static/css/main.css': ['static/css/main.less']
         }
       }
     },
@@ -161,7 +131,7 @@ module.exports = function(grunt) {
       main: {
         // Prefix `static/css/main.css` and overwrite.
         expand: true,
-        src: ['static/css/main.css']
+        src: ['<%= loc.dist %>/static/css/main.css']
       },
     },
 
@@ -175,8 +145,8 @@ module.exports = function(grunt) {
      */
     browserify: {
       build: {
-        src: 'src/static/js/routes/**/*.js',
-        dest: 'static/js/routes/common.min.js'
+        src: '<%= loc.src %>/static/js/routes/**/*.js',
+        dest: '<%= loc.dist %>/static/js/routes/common.min.js'
       },
       options: {
         // Note: The transforms for minification and
@@ -192,20 +162,20 @@ module.exports = function(grunt) {
      */
     banner:
       '/*!\n' +
-      ' *              ad$$             $$\n' +
-      ' *             d$"               $$\n' +
-      ' *             $$                $$\n' +
-      ' *   ,adPYba,  $$$$$ $b,dPYba,   $$,dPYba,\n' +
-      ' *  aP\'    \'$: $$    $$P\'   \'$a  $$P\'   \'$a\n' +
-      ' *  $(         $$    $$(     )$  $$(     )$\n' +
-      ' *  "b,    ,$: $$    $$b,   ,$"  $$b,   ,$"\n' +
-      ' *   `"Ybd$"\'  $$    $$`YbdP"\'   $$`Ybd$"\'\n' +
-      ' *                   $$\n' +
-      ' *                   $$\n' +
-      ' *                   ""\n' +
+      ' *               ad$$               $$\n' +
+      ' *              d$"                 $$\n' +
+      ' *              $$                  $$\n' +
+      ' *   ,adPYba.   $$$$$  $$.,dPYba.   $$.,dPYba.\n' +
+      ' *  aP′    `$:  $$     $$P′    `$a  $$P′    `$a\n' +
+      ' *  $(          $$     $$(      )$  $$(      )$\n' +
+      ' *  "b.    ,$:  $$     $$b.    ,$"  $$b.    ,$"\n' +
+      ' *   `"Ybd$"′   $$     $$`"YbdP"′   $$`"YbdP"′\n' +
+      ' *                     $$\n' +
+      ' *                     $$\n' +
+      ' *                     $$\n' +
       ' *\n' +
       ' *  <%= pkg.name %> - v<%= pkg.version %>\n' +
-      ' *  <%= pkg.homepage %>' +
+      ' *  <%= pkg.homepage %>\n' +
       ' *  A public domain work of the Consumer Financial Protection Bureau\n' +
       ' */',
 
@@ -217,7 +187,7 @@ module.exports = function(grunt) {
           linebreak: true
         },
         files: {
-          src: ['static/css/*.min.css']
+          src: ['<%= loc.dist %>/static/css/*.min.css']
         }
       },
       js: {
@@ -227,7 +197,7 @@ module.exports = function(grunt) {
           linebreak: true
         },
         files: {
-          src: ['static/js/**/*.min.js']
+          src: ['<%= loc.dist %>/static/js/**/*.min.js']
         }
       }
     },
@@ -241,12 +211,12 @@ module.exports = function(grunt) {
       'ie-alternate': {
         options: {
           // Flatten all media queries with a min-width over 960 or lower.
-          // All media queries over 960 will be excluded fromt he stylesheet.
+          // All media queries over 960 will be excluded from the stylesheet.
           // EM calculation: 960 / 16 = 60
           legacyWidth: 60
         },
         files: {
-          'static/css/main.ie.css': 'static/css/main.css'
+          '<%= loc.dist %>/static/css/main.ie.css': '<%= loc.dist %>/static/css/main.css'
         }
       }
     },
@@ -302,8 +272,8 @@ module.exports = function(grunt) {
             quiet: envQuiet
         },
         src: [
-            'src/static/js/**/*.js',
-            '_tests/unit_tests/*.js'
+            '<%= loc.src %>/static/js/**/*.js',
+            '<%= loc.test %>/unit_tests/*.js'
         ]
       }
     },
@@ -313,10 +283,10 @@ module.exports = function(grunt) {
      */
     mocha_istanbul: {
       coverage: {
-        src: ['_tests/unit_tests/**/*.js'],
+        src: ['<%= loc.test %>/unit_tests/**/*.js'],
         options: {
-          coverageFolder: '_tests/unit_test_coverage',
-          excludes: ['_tests/unit_tests/**/*.js']
+          coverageFolder: '<%= loc.test %>/unit_test_coverage',
+          excludes: ['<%= loc.test %>/unit_tests/**/*.js']
         }
       }
     },
@@ -329,7 +299,7 @@ module.exports = function(grunt) {
      */
     watch: {
       all: {
-        files: ['static/css/*.less', 'src/static/js/**/*.js', 'Gruntfile.js'],
+        files: ['static/css/*.less', '<%= loc.src %>/static/js/**/*.js', 'Gruntfile.js'],
         tasks: ['default']
       },
       css: {
@@ -337,7 +307,7 @@ module.exports = function(grunt) {
         tasks: ['cssdev']
       },
       cssjs: {
-        files: ['static/css/*.less', 'src/static/js/**/*.js'],
+        files: ['static/css/*.less', '<%= loc.src %>/static/js/**/*.js'],
         tasks: ['cssdev', 'jsdev']
       }
     },
@@ -424,7 +394,7 @@ module.exports = function(grunt) {
   /**
    * Create custom task aliases and combinations.
    */
-  grunt.registerTask('vendor', ['bower:install', 'string-replace:chosen', 'string-replace:static-legacy',
+  grunt.registerTask('vendor', ['bower:install', 'string-replace:chosen',
                                 'copy:static-legacy', 'concat:cf-less']);
   grunt.registerTask('cssdev', ['less', 'autoprefixer', 'legacssy', 'usebanner:css']);
   grunt.registerTask('jsdev', ['lintjs', 'browserify:build', 'usebanner:js']);